# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BearerAuthError": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "default": "missing or invalid bearer token",
            "title": "Detail",
            "type": "string"
          }
        },
        "title": "BearerAuthError",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchEntry": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "entry_id": {
            "title": "Entry Id",
            "type": "string"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folder_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Path"
          },
          "ingest_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingest Status"
          },
          "lifecycle": {
            "title": "Lifecycle",
            "type": "string"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "related_entries": {
            "items": {
              "$ref": "#/components/schemas/SearchRelatedEntry"
            },
            "title": "Related Entries",
            "type": "array"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "entry_id",
          "display_name",
          "folder_id",
          "folder_path",
          "lifecycle",
          "mime_type",
          "size_bytes",
          "ingest_status",
          "created_at",
          "updated_at",
          "related_entries"
        ],
        "title": "SearchEntry",
        "type": "object"
      },
      "SearchRelatedEntry": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "entry_id": {
            "title": "Entry Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "entry_id",
          "display_name",
          "score"
        ],
        "title": "SearchRelatedEntry",
        "type": "object"
      },
      "SearchResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/SearchEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "q": {
            "title": "Q",
            "type": "string"
          }
        },
        "required": [
          "q",
          "count",
          "entries"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Library",
    "version": "0.3.6"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/discover/{entry_id}": {
      "get": {
        "description": "Random-walk recommendation from a seed entry. Drives the\n`/discover` REPL command and the related_entries pre-fill in\nsearch/get_metadata. Vetted edges only by default; pass\ninclude_unvetted=true to walk the raw graph. Pass vet=true to queue\nbackground vetting for the seed's direct raw edges; the response itself\nstays pure-read.",
        "operationId": "discover_v1_discover__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "top_k",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 30,
              "minimum": 1,
              "title": "Top K",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_unvetted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unvetted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "vet",
            "required": false,
            "schema": {
              "default": false,
              "title": "Vet",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Discover V1 Discover  Entry Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover",
        "tags": [
          "user_files"
        ]
      }
    },
...
    "/v1/file-entries/{entry_id}/content": {
      "get": {
        "description": "Inline-disposition variant of `/download` — used by the viewer\niframe so PDFs and images render in the browser instead of getting\nsaved to disk. The `/download` endpoint still exists for the\nexplicit Download button (forces save-as).",
        "operationId": "file_entry_content_v1_file_entries__entry_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "File Entry Content",
        "tags": [
          "user_files"
        ]
      }
    },
    "/v1/file-entries/{entry_id}/download": {
      "get": {
        "operationId": "file_entry_download_v1_file_entries__entry_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "File Entry Download",
        "tags": [
          "user_files"
        ]
      }
    },
...
    "/v1/file-entries/{entry_id}/metadata": {
      "get": {
        "operationId": "file_entry_metadata_v1_file_entries__entry_id__metadata_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response File Entry Metadata V1 File Entries  Entry Id  Metadata Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "File Entry Metadata",
        "tags": [
          "user_files"
        ]
      }
    },
...
    "/v1/file-entries/{entry_id}/path": {
      "get": {
        "description": "Folder ancestor chain (root → leaf) for an entry.\n\nThe GUI calls this when the user clicks a search hit or a\nchat citation, so the Library tree can expand each ancestor in\norder before selecting the file. Returns 404 if the entry is\nsoft-deleted or unknown.",
        "operationId": "file_entry_path_v1_file_entries__entry_id__path_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response File Entry Path V1 File Entries  Entry Id  Path Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "File Entry Path",
        "tags": [
          "user_files"
        ]
      }
    },
    "/v1/file-entries/{entry_id}/preview-text": {
      "get": {
        "operationId": "file_entry_preview_text_v1_file_entries__entry_id__preview_text_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_chars",
            "required": false,
            "schema": {
              "default": 2000000,
              "maximum": 5000000,
              "minimum": 1,
              "title": "Max Chars",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response File Entry Preview Text V1 File Entries  Entry Id  Preview Text Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "File Entry Preview Text",
        "tags": [
          "user_files"
        ]
      }
    },
...
    "/v1/folders/{folder_id}/download": {
      "get": {
        "operationId": "folder_download_v1_folders__folder_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder_id",
            "required": true,
            "schema": {
              "title": "Folder Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Folder Download",
        "tags": [
          "user_files"
        ]
      }
    },
...
    "/v1/search": {
      "get": {
        "operationId": "search_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BearerAuthError"
                }
              }
            },
            "description": "Missing or invalid bearer token. Only raised when LIBRARY_API_TOKEN is configured. Tokenless loopback is the default.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "example": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "user_files"
        ]
      }
    },
...
  }
}